+++ xyzpy/gen/batch.py
@@ -1254,8 +1254,14 @@ def gen_qsub_script(
         opts['batch_ids'] = batch_ids
 
     script += _BASE_QSUB_SCRIPT_END
+    script = script.format(**opts)
 
-    return script.format(**opts)
+    if scheduler == 'pbs' and len(batch_ids) == 1:
+        # PBS can't handle arrays jobs of size 1...
+        script = (script.replace('#PBS -J 1-1\n', "")
+                        .replace("$PBS_ARRAY_INDEX", '1'))
+
+    return script
 
 
 def qsub_grow(