+++ docs/changelog.rst
@@ -12,6 +12,7 @@ v0.3.2 (Unreleased)
 **Enhancements**
 
 - Add PBS support to :meth:`xyzpy.Crop.qsub_grow` for distributed crop growing
+- Fix PBS crop submission for job arrays of size 1
 - Add :func:`xyzpy.save_merge_ds` for manually aggregating datasets to disk
 - Add ``allow_incomplete=True`` option to :meth:`xyzpy.Crop.reap` for gathering data even if the crop is not fully grown (:issue:`7` )
 - Make new :class:`~xyzpy.Crop` instances by default automatically load information from disk if they have been already prepared/sown (:issue:`7` )